wikimedia-m1/requirements.txt
confluent-kafka
psycopg[binary]==3.3.3
requests==2.32.3
python-dateutil==2.9.0.post0
